Rewrite UI for JavaFX / Convert to maven build
