+++ .gitignore
@@ -1,4 +1,8 @@
 /bin/
-
-.settings/
+/lib/
+/test/
 /output/
+/target/
+/.metadata/
+.settings/
+
