+++ src/java/com/javadeobfuscator/deobfuscator/ui/util/InvalidJarException.java
@@ -1,0 +1,10 @@
+package com.javadeobfuscator.deobfuscator.ui.util;
+
+/**
+ * Exception to be thrown when a jar has been loaded for the deobfuscator
+ * wrapper, but jar was not an instance of JavaDeobfuscator.
+ */
+@SuppressWarnings("serial")
+public class InvalidJarException extends Exception {
+
+}
