+++ src/resources/META-INF/services/com.fasterxml.jackson.core.JsonFactory
@@ -1,0 +1,1 @@
+com.fasterxml.jackson.core.JsonFactory
