+++ src/resources/META-INF/services/com.fasterxml.jackson.core.ObjectCodec
@@ -1,0 +1,1 @@
+com.fasterxml.jackson.databind.ObjectMapper
